A simple nose under player
I added a simple nose under the player to show where the player is facing.
Sadly, I was trying to attach it under the camera, but it won't render with network.
One way to achieve this is to write a script, but I don't want to use any resource just to update the position of a nose, we should discuss a little bit about this.