Cleaned up spawn location, rotation
Tweaked spawn location so that you spawn in the center of your spawn platform.

Also added code so that all players spawn facing the inside wall where the instructions are.

The updated spawn is a prefab named "Spawn 2 1". You can find it inside Prefabs/Level Prefabs/Components.

To add it to the level, drag the prefab into the scene, delete/hide the old spawn, then update the network manager with the new spawn locations.